requirements.txt
# === Core API ===
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
pydantic>=2.0

# === LangChain / LangGraph ===
langchain>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.28

# === Vector DB ===
qdrant-client>=1.11.0

# === VLM (Qwen2-VL) ===
transformers>=4.45.0
torch>=2.4.0
torchvision>=0.19.0
accelerate>=0.26.0
Pillow>=10.4.0
qwen-vl-utils>=0.0.8

# === RAG Embedding ===
sentence-transformers>=3.0.0

# === Experiment Tracking ===
mlflow>=2.16.0
pandas>=2.2.0

# === Utilities ===
python-dotenv>=1.0.1
httpx>=0.27.0